pyproject.toml
[tool.poetry]
name = "adminpanel"
version = "0.1.0"
description = ""
authors = ["SowaSova <cei4bio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
psycopg2 = "^2.9.10"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"
whitenoise = "^6.8.2"
pillow = "^11.0.0"
django-storages = "^1.14.4"
boto3 = "^1.35.55"
django-import-export = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
